+++ app/src/main/java/com/gmail/eamosse/imdb/ui/home/adapter/MovieHandler.kt
@@ -1,5 +1,5 @@
 package com.gmail.eamosse.imdb.ui.home.adapter
 
 interface MovieHandler {
-    fun onShowMovieDetails(id: Int)
+    fun onShowMovieDetails(id: Long)
 }
